do typescript
