+++ context.js
@@ -1,4 +1,1 @@
-const context = require('./lib/context');
-module.exports = {
-  getContext: context.getContext,
-};
+module.exports = require('./dist/context');
