+++ lib/babelTransformContext.js
@@ -1,6 +1,6 @@
 const { annotateAsPure } = require('./astUtils');
 
-const IMPORT_PATH = 'next-rpc/lib/context';
+const IMPORT_PATH = 'next-rpc/dist/context-internal';
 
 /**
  * @typedef {{ apiDir: string, isServer: boolean }} PluginOptions
