+++ lib/babelTransformRpc.js
@@ -1,7 +1,7 @@
 const { annotateAsPure } = require('./astUtils');
 
-const IMPORT_PATH_SERVER = 'next-rpc/lib/server';
-const IMPORT_PATH_BROWSER = 'next-rpc/lib/browser';
+const IMPORT_PATH_SERVER = 'next-rpc/dist/server';
+const IMPORT_PATH_BROWSER = 'next-rpc/dist/browser';
 
 /**
  * @param {typeof import('@babel/core').types} t
