+++ package.json
@@ -2,12 +2,13 @@
   "name": "next-rpc",
   "version": "3.1.6",
   "description": "Call serverside code from the browser in next.js applications",
-  "main": "lib",
+  "main": "dist/index.js",
   "repository": "https://github.com/Janpot/next-rpc",
   "author": "Jan Potoms",
   "license": "MIT",
   "files": [
     "lib/*",
+    "dist/*",
     "context.js",
     "context.d.ts"
   ],
@@ -36,13 +37,14 @@
   },
   "scripts": {
     "prettier": "prettier --check \"**/*.{js,ts,jsx,tsx,json,yml,md}\"",
-    "typescript": "tsc",
-    "lint": "$npm_execpath run prettier && $npm_execpath run typescript",
+    "lint": "$npm_execpath run prettier",
     "jest": "jest --testRunner='jest-circus/runner' --runInBand",
     "test": "$npm_execpath run lint && $npm_execpath run jest --runInBand",
     "fix": "$npm_execpath run prettier -- --write",
     "next": "next",
-    "prepublishOnly": "$npm_execpath run test"
+    "prepublishOnly": "$npm_execpath run build && $npm_execpath run test",
+    "build": "tsc",
+    "watch": "tsc --watch"
   },
   "peerDependencies": {
     "next": ">=10"
